requirements.txt
paho-mqtt==1.5.1
pymodbus==2.5.1
pyserial==3.5
six==1.15.0
